add vllm test
